# Salus/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salus API",
        "description": "Documentation de l'API Salus.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/doctors/{id}": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Voir un medecin",
                "operationId": "cadaed8345b4446c68eb692a3430a5a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail du medecin",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail du medecin"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Doctor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Medecin introuvable"
                    }
                }
            }
        },
        "/api/doctors/search": {
            "get": {
                "tags": [
                    "Doctors"
                ],
                "summary": "Rechercher des medecins",
                "operationId": "e5fdbd636c63b391aec3ce0af32c6633",
                "parameters": [
                    {
                        "name": "specialty",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Cardiologie"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Casablanca"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultats de recherche",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Resultats de recherche"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Doctor"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Doctor": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Dr. Amina El Idrissi"
                    },
                    "specialty": {
                        "type": "string",
                        "example": "Cardiologie"
                    },
                    "city": {
                        "type": "string",
                        "example": "Casablanca"
                    },
                    "consultation_price": {
                        "type": "string",
                        "example": "350.00"
                    },
                    "available_days": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Monday"
                        }
                    },
                    "yearsofexperience": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Entrez le token au format: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}